implementing settings in dropdown menu
