+++ components/admin/AdminShell.tsx
@@ -37,7 +37,7 @@ import {
   IconStatusChange,
   IconTicket,
 } from '@tabler/icons-react';
-import UserProfile from '@/components/UserAvatar';
+import UserAvatar from '@/components/UserAvatar';
 import { Session } from 'next-auth';
 import { useDisclosure } from '@mantine/hooks';
 import { Competition } from '@prisma/client';
@@ -45,7 +45,7 @@ import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import useSWR from 'swr';
 import superjson from 'superjson';
-import SettingsModal from '../settings/SettingsModal';
+import { useSession } from 'next-auth/react';
 
 interface Props {
   session: Session;
@@ -92,7 +92,6 @@ export default function AdminShell({
     <CompetitionContext.Provider
       value={{ competition: competitions?.find((c) => c.code === comp) }}
     >
-      <SettingsModal opened={true} />
       <AppShell
         header={{ height: 60 }}
         navbar={{
@@ -128,7 +127,7 @@ export default function AdminShell({
               </Title>
             </Group>
 
-            <UserProfile session={session} />
+            <UserAvatar session={session!} />
           </Group>
         </AppShellHeader>
